change round seconds to u16

+++ ferrogallic_shared/src/api/game.rs
@@ -51,7 +51,7 @@ pub struct GameState {
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct GameConfig {
     pub rounds: u8,
-    pub guess_seconds: u8,
+    pub guess_seconds: u16,
 }
 
 impl Default for GameConfig {
